serperf: Add signal_handler and exit_print()
Add signal handler so we can check how many msgs the server has sent and
received. Also implemented exit_print(), so we can show the number of
msgs and the time that the tool has been running in every case, i.e. if
it times out, we can see now where in the process it has stopped.
Redefine time measuring variables as globals so we can show them in
every exit scenario.

Signed-off-by: Hernán Gonzalez <hernan@vanguardiasur.com.ar>